reg: removed compiler warnings

+++ cob_3d_registration/common/include/cob_3d_registration/features/edges.h
@@ -50,7 +50,7 @@ public:
 
   virtual bool hidden_build() {
 
-    ROS_INFO("calc edges for source %d", this->org_in_.size());
+    ROS_INFO("calc edges for source %d", (int)this->org_in_.size());
     {
       /*extractFeatures(this->org_in_, feature_in_);
       this->org_in_.points.clear();
@@ -67,11 +67,11 @@ public:
       feature_in_=this->org_in_;
     }
 
-    ROS_INFO("calc edges for target %d", this->org_out_.size());
+    ROS_INFO("calc edges for target %d", (int)this->org_out_.size());
     {
       extractFeatures(this->org_out_, feature_out_);
       this->org_out_.points.clear();
-      for(int i=0; i<feature_out_.size(); i++) {
+      for(int i=0; i<(int)feature_out_.size(); i++) {
         Point p;
         memset(&p,0,sizeof(Point));
         p.x=feature_out_.points[i].x;
@@ -84,7 +84,7 @@ public:
     }
 
     //need to rebuild tree
-    ROS_INFO("build tree %d %d", this->org_in_.size(), this->org_out_.size());
+    ROS_INFO("build tree %d %d", (int)this->org_in_.size(), (int)this->org_out_.size());
     this->tree_.reset (new pcl::KdTreeFLANN<Point>);
     if(this->org_in_.size()>0) this->tree_->setInputCloud(this->org_in_.makeShared());
     ROS_INFO("build tree done");
